+++ provider/provider.go
@@ -5,6 +5,7 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/cluster"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/index"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ingest"
+	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/logstash"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/security"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -128,6 +129,7 @@ func New(version string) func() *schema.Provider {
 				"elasticstack_elasticsearch_index_lifecycle":       index.ResourceIlm(),
 				"elasticstack_elasticsearch_index_template":        index.ResourceTemplate(),
 				"elasticstack_elasticsearch_ingest_pipeline":       ingest.ResourceIngestPipeline(),
+				"elasticstack_elasticsearch_logstash_pipeline":     logstash.ResourceLogstashPipeline(),
 				"elasticstack_elasticsearch_security_role":         security.ResourceRole(),
 				"elasticstack_elasticsearch_security_role_mapping": security.ResourceRoleMapping(),
 				"elasticstack_elasticsearch_security_user":         security.ResourceUser(),